+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - H2 adding parentheses to CQL WHERE clauses and not appending ALLOW FILTERING.
+- Fix `ModelType.withinErrorBound()` never returning true with error bound zero.
 
 ### Security